requirements.txt
streamlit
numpy
requests
json
Pillow
